Рефакторинг: исправление, изменение содержания модулей
